Add a button that allows user to create new grid
Commented out loop that created a 16 x 16 grid for testing. New button
has an event listener that will prompt the user for squares per side.
Wrote a new function that creates the new grid with a separate callback
function for calculating the new square sizes. Moved the event listeners
for drawing over squares into new grid function so that the new grid is
now responsive.