Add getNumberOfSquares function
New function gets input from user and checks to make sure input isn't
over 100 and sets input to default size of 16 if so for simplicity.
Some slight tweaks are also included for width but I am researching how
to make sure children fill the parent container for flexbox as it's not
completely filling with an input of 6 so these can be disregarded.